simos.py -> oss subpackage.
