+++ angr/misc/__init__.py
@@ -2,3 +2,4 @@ from . import ux
 from . import autoimport
 from .loggers import Loggers
 from . import graph
+from .range import IRange
