+++ angr/misc/range.py
@@ -1,0 +1,21 @@
+
+class IRange(object):
+    """
+    A simple range object for testing inclusion. Like xrange but works for huge numbers.
+    """
+    __slots__ = ('start', 'end')
+
+    def __init__(self, start, end):
+        self.start = start
+        self.end = end
+
+    def __contains__(self, k):
+        if type(k) in (int, long):
+            return k >= self.start and k < self.end
+        return False
+
+    def __getstate__(self):
+        return self.start, self.end
+
+    def __setstate__(self, state):
+        self.start, self.end = state
